Seperation of Concerns + fix profile_widget
